Dropped the BoxedIterator
It was useless for the intended situation: unpacking a Iterator<Result<V,E>> into a Result<BoxedIterator<V>,E> would have to walk the entire iterator anyway. So why not collect into a Result<Vec<V>, E>. In other places a Box<Iterator<T>> was fine.
In the mean time this BoxedIterator was giving OOM and SEGV problems with debug build (but not the release build, funny bug in the rust compiler?), see http://stackoverflow.com/q/38254647/859279.