Do not color comparisons that changed less than 3%
Currently, comparisons are always color coded: any improvement is
shown in green and any regression is shown in red. It doesn't matter
if the change is 0.78% or if it's 7.8%, the output looks the same.

This change introduces a 3% cut-off so that comparison where the
change is smaller than 3% are shown without color. The value of 3% is
pretty arbitrary and should perhaps be set lower.

The cut-off is not connected with the --threshold option, so it's
possible to ask for a threshold of 1% and still see uncolored lines in
the output. That may or may not be a feature.